+++ ext/wankel/wankel_stream_encoder.h
@@ -1,5 +1,5 @@
-#ifndef WANKEL_SAX_ENCODER
-#define WANKEL_SAX_ENCODER
+#ifndef WANKEL_STREAM_ENCODER
+#define WANKEL_STREAM_ENCODER
 
 #include <ruby.h>
 #include <ruby/encoding.h>
@@ -8,6 +8,6 @@
 
 #include "yajl_helpers.h"
 
-void Init_wankel_sax_encoder();
+void Init_wankel_stream_encoder();
 
 #endif