+++ ext/wankel/wankel_stream_parser.c
@@ -1,6 +1,6 @@
-#include "wankel_sax_parser.h"
+#include "wankel_stream_parser.h"
 
-static VALUE sax_parser_initialize(int argc, VALUE * argv, VALUE self);
+static VALUE stream_parser_initialize(int argc, VALUE * argv, VALUE self);
 
 static ID  sym_read_buffer_size, sym_symbolize_keys;
 
@@ -10,30 +10,30 @@ static ID  intern_on_null, intern_on_boolean, intern_on_integer,
     intern_on_double, intern_on_string, intern_on_map_start, intern_on_map_key,
     intern_on_map_end, intern_on_array_start, intern_on_array_end;
 	
-static VALUE c_wankel, c_wankelParser, c_saxParser, e_parseError, e_encodeError;
+static VALUE c_wankel, c_wankelParser, c_streamParser, e_parseError, e_encodeError;
 		   
 // Callbacks =================================================================
-yajl_callbacks sax_parser_callbacks(VALUE self);
-int sax_parser_callback_on_null(void *ctx);
-int sax_parser_callback_on_boolean(void *ctx, int boolVal);
-int sax_parser_callback_on_number(void *ctx, const char * numberVal, size_t numberLen);
-int sax_parser_callback_on_string(void *ctx, const unsigned char * stringVal, size_t stringLen);
-int sax_parser_callback_on_map_start(void *ctx);
-int sax_parser_callback_on_map_key(void *ctx, const unsigned char * key, size_t keyLen);
-int sax_parser_callback_on_map_end(void *ctx);
-int sax_parser_callback_on_array_start(void *ctx);
-int sax_parser_callback_on_array_end(void *ctx);
+yajl_callbacks stream_parser_callbacks(VALUE self);
+int stream_parser_callback_on_null(void *ctx);
+int stream_parser_callback_on_boolean(void *ctx, int boolVal);
+int stream_parser_callback_on_number(void *ctx, const char * numberVal, size_t numberLen);
+int stream_parser_callback_on_string(void *ctx, const unsigned char * stringVal, size_t stringLen);
+int stream_parser_callback_on_map_start(void *ctx);
+int stream_parser_callback_on_map_key(void *ctx, const unsigned char * key, size_t keyLen);
+int stream_parser_callback_on_map_end(void *ctx);
+int stream_parser_callback_on_array_start(void *ctx);
+int stream_parser_callback_on_array_end(void *ctx);
 
 // Ruby GC ===================================================================
-VALUE sax_parser_alloc(VALUE);
-//void sax_parser_mark(void *);
-void sax_parser_free(void * parser);
+VALUE stream_parser_alloc(VALUE);
+//void stream_parser_mark(void *);
+void stream_parser_free(void * parser);
 
-VALUE sax_parser_initialize(int argc, VALUE * argv, VALUE self) {
+VALUE stream_parser_initialize(int argc, VALUE * argv, VALUE self) {
     VALUE defaults = rb_const_get(c_wankel, intern_DEFAULTS);
     VALUE klass = rb_funcall(self, rb_intern("class"), 0);
     VALUE options, rbufsize;
-    sax_parser * p;
+    stream_parser * p;
 
     rb_scan_args(argc, argv, "01", &options);
     if (rb_const_defined(klass, intern_DEFAULTS)) { 
@@ -47,8 +47,8 @@ VALUE sax_parser_initialize(int argc, VALUE * argv, VALUE self) {
     }
     options = rb_iv_get(self, "@options");
 
-    Data_Get_Struct(self, sax_parser, p);
-    p->callbacks = sax_parser_callbacks(self);
+    Data_Get_Struct(self, stream_parser, p);
+    p->callbacks = stream_parser_callbacks(self);
     p->alloc_funcs.malloc = yajl_helper_malloc;
     p->alloc_funcs.realloc = yajl_helper_realloc;
     p->alloc_funcs.free = yajl_helper_free;
@@ -69,13 +69,13 @@ VALUE sax_parser_initialize(int argc, VALUE * argv, VALUE self) {
     return self;
 }
 
-static VALUE sax_parser_parse(int argc, VALUE * argv, VALUE self) {
+static VALUE stream_parser_parse(int argc, VALUE * argv, VALUE self) {
     const char * cptr;
     size_t len;
     yajl_status status;
-    sax_parser * p;
+    stream_parser * p;
     VALUE input;
-    Data_Get_Struct(self, sax_parser, p);
+    Data_Get_Struct(self, stream_parser, p);
 	
     rb_scan_args(argc, argv, "10", &input);
     if (TYPE(input) == T_STRING) {
@@ -101,12 +101,12 @@ static VALUE sax_parser_parse(int argc, VALUE * argv, VALUE self) {
     return Qnil;
 }
 
-static VALUE sax_parser_write(VALUE self, VALUE input) {
+static VALUE stream_parser_write(VALUE self, VALUE input) {
     const char * cptr;
     size_t len;
     yajl_status status;
-    sax_parser * p;
-    Data_Get_Struct(self, sax_parser, p);
+    stream_parser * p;
+    Data_Get_Struct(self, stream_parser, p);
     
     Check_Type(input, T_STRING);
     cptr = RSTRING_PTR(input);
@@ -117,10 +117,10 @@ static VALUE sax_parser_write(VALUE self, VALUE input) {
     return Qnil;
 }
 
-static VALUE sax_parser_complete(VALUE self) {
+static VALUE stream_parser_complete(VALUE self) {
     yajl_status status;
-    sax_parser * p;
-    Data_Get_Struct(self, sax_parser, p);
+    stream_parser * p;
+    Data_Get_Struct(self, stream_parser, p);
     
 
     status = yajl_complete_parse(p->h);
@@ -129,20 +129,20 @@ static VALUE sax_parser_complete(VALUE self) {
     return Qnil;
 }
 
-void Init_wankel_sax_parser() {
+void Init_wankel_stream_parser() {
     c_wankel = rb_const_get(rb_cObject, rb_intern("Wankel"));
     c_wankelParser = rb_const_get(c_wankel, rb_intern("Parser"));
-    c_saxParser = rb_define_class_under(c_wankel, "SaxParser", rb_cObject);
+    c_streamParser = rb_define_class_under(c_wankel, "StreamParser", rb_cObject);
     e_parseError = rb_const_get(c_wankel, rb_intern("ParseError"));
     e_encodeError = rb_const_get(c_wankel, rb_intern("EncodeError"));
     
-    rb_define_alloc_func(c_saxParser, sax_parser_alloc);
-    rb_define_method(c_saxParser, "initialize", sax_parser_initialize, -1);
-    rb_define_method(c_saxParser, "parse", sax_parser_parse, -1);
-    rb_define_method(c_saxParser, "write", sax_parser_write, 1);
-    rb_define_method(c_saxParser, "complete", sax_parser_complete, 0);
+    rb_define_alloc_func(c_streamParser, stream_parser_alloc);
+    rb_define_method(c_streamParser, "initialize", stream_parser_initialize, -1);
+    rb_define_method(c_streamParser, "parse", stream_parser_parse, -1);
+    rb_define_method(c_streamParser, "write", stream_parser_write, 1);
+    rb_define_method(c_streamParser, "complete", stream_parser_complete, 0);
     
-    rb_define_alias(c_saxParser, "<<", "write");
+    rb_define_alias(c_streamParser, "<<", "write");
     
     intern_merge = rb_intern("merge");
     intern_clone = rb_intern("clone");
@@ -164,52 +164,52 @@ void Init_wankel_sax_parser() {
 }
 
 // Callbacks =================================================================
-yajl_callbacks sax_parser_callbacks(VALUE self) {
+yajl_callbacks stream_parser_callbacks(VALUE self) {
     yajl_callbacks callbacks = { NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL };
 
     if(rb_respond_to(self, intern_on_null)) {
-        callbacks.yajl_null = sax_parser_callback_on_null;
+        callbacks.yajl_null = stream_parser_callback_on_null;
     }
     if(rb_respond_to(self, intern_on_boolean)) {
-        callbacks.yajl_boolean = sax_parser_callback_on_boolean;
+        callbacks.yajl_boolean = stream_parser_callback_on_boolean;
     }
     if(rb_respond_to(self, intern_on_integer) || rb_respond_to(self, intern_on_double)) {
-        callbacks.yajl_number = sax_parser_callback_on_number;
+        callbacks.yajl_number = stream_parser_callback_on_number;
     }
     if(rb_respond_to(self, intern_on_string)) {
-        callbacks.yajl_string = sax_parser_callback_on_string;
+        callbacks.yajl_string = stream_parser_callback_on_string;
     }
 
     if(rb_respond_to(self, intern_on_map_start)) {
-        callbacks.yajl_start_map = sax_parser_callback_on_map_start;
+        callbacks.yajl_start_map = stream_parser_callback_on_map_start;
     }
     if(rb_respond_to(self, intern_on_map_key)) {
-        callbacks.yajl_map_key = sax_parser_callback_on_map_key;
+        callbacks.yajl_map_key = stream_parser_callback_on_map_key;
     }
     if(rb_respond_to(self, intern_on_map_end)) {
-        callbacks.yajl_end_map = sax_parser_callback_on_map_end;
+        callbacks.yajl_end_map = stream_parser_callback_on_map_end;
      }
     if(rb_respond_to(self, intern_on_array_start)) {
-        callbacks.yajl_start_array = sax_parser_callback_on_array_start;
+        callbacks.yajl_start_array = stream_parser_callback_on_array_start;
     }
     if(rb_respond_to(self, intern_on_array_end)) {
-        callbacks.yajl_end_array = sax_parser_callback_on_array_end;
+        callbacks.yajl_end_array = stream_parser_callback_on_array_end;
     }
 
     return callbacks;
 }
 
-int sax_parser_callback_on_null(void *ctx) {
+int stream_parser_callback_on_null(void *ctx) {
     rb_funcall((VALUE)ctx, intern_on_null, 0);
     return 1;
 }
 
-int sax_parser_callback_on_boolean(void *ctx, int boolVal) {
+int stream_parser_callback_on_boolean(void *ctx, int boolVal) {
     rb_funcall((VALUE)ctx, intern_on_boolean, 1, (boolVal ? Qtrue : Qfalse));
     return 1;
 }
 
-int sax_parser_callback_on_number(void *ctx, const char * numberVal, size_t numberLen) {
+int stream_parser_callback_on_number(void *ctx, const char * numberVal, size_t numberLen) {
 	char buf[numberLen+1];
 	buf[numberLen] = 0;
 	memcpy(buf, numberVal, numberLen);
@@ -227,38 +227,38 @@ int sax_parser_callback_on_number(void *ctx, const char * numberVal, size_t numb
 	return 1;
 }
 
-int sax_parser_callback_on_string(void *ctx, const unsigned char * stringVal, size_t stringLen) {
+int stream_parser_callback_on_string(void *ctx, const unsigned char * stringVal, size_t stringLen) {
     rb_funcall((VALUE)ctx, intern_on_string, 1, rb_str_new((const char *) stringVal, stringLen));
     return 1;
 }
-int sax_parser_callback_on_map_start(void *ctx) {
+int stream_parser_callback_on_map_start(void *ctx) {
     rb_funcall((VALUE)ctx, intern_on_map_start, 0);
     return 1;
 }
-int sax_parser_callback_on_map_key(void *ctx, const unsigned char * key, size_t keyLen) {
+int stream_parser_callback_on_map_key(void *ctx, const unsigned char * key, size_t keyLen) {
     rb_funcall((VALUE)ctx, intern_on_map_key, 1, rb_str_new((const char *)key, keyLen));
     return 1;
 }
-int sax_parser_callback_on_map_end(void *ctx) {
+int stream_parser_callback_on_map_end(void *ctx) {
     rb_funcall((VALUE)ctx, intern_on_map_end, 0);
     return 1;
 }
-int sax_parser_callback_on_array_start(void *ctx) {
+int stream_parser_callback_on_array_start(void *ctx) {
     rb_funcall((VALUE)ctx, intern_on_array_start, 0);
     return 1;
 }
-int sax_parser_callback_on_array_end(void *ctx) {
+int stream_parser_callback_on_array_end(void *ctx) {
     rb_funcall((VALUE)ctx, intern_on_array_end, 0);
     return 1;
 }
 
 // Ruby GC ===================================================================
-VALUE sax_parser_alloc(VALUE klass) {
-    sax_parser * p;
-    return Data_Make_Struct(klass, sax_parser, 0, sax_parser_free, p);
+VALUE stream_parser_alloc(VALUE klass) {
+    stream_parser * p;
+    return Data_Make_Struct(klass, stream_parser, 0, stream_parser_free, p);
 }
 
-void sax_parser_free(void * handle) {
-    sax_parser * p = handle;
+void stream_parser_free(void * handle) {
+    stream_parser * p = handle;
 	yajl_free(p->h);
 }